shutdown handler
Reset SC3900 lights to the default settings (i.e. as how they are when
we connect the SC3900 for the first time in normal MIDI mode) when
leaving mixxx.

Unfortunately, we can't grab the connected SC3900 decks channel number
on shutdown, so we blindly assume that all the mixxx decks are SC3900,
and we send the MIDI messages to all these decks.