do not manage the VINYL led
Let the SC3900 unit handle the VINYL led status.
In normal MIDI mode, the led is always on even if we press the VINYL
button. The vinyl disc acts like a jog wheel.

In hybrid MIDI mode, the vinyl mode can be toggled, and the led is
switched ON/OFF by the unit itself.
The vinyl disc can act like a jogwheel (VINYL mode OFF) or like a vinyl
disc (VINYL mode ON, with rotating platter).
In hybrid MIDI mode, the vinyl disc and VINYL buttons aren't sending any
MIDI messages. Instead, they influence on the audio signal emitted by the
SC3900 signal for DVS.