fix pitch fader precision
The 438a05886af466476c73a7feb615b6e252640e42 commit which fixed the
warning messages of mixxx when moving the pitch fader has introduced a
loss of precision in the computing of the pitch rate.
Correctly handling the MSB and LSB data allow us to keep high precision
and to avoid warning messages.