fix CDJ like CUE button hold
When paused on a CUE point, holding down the CUE button should play
the track untill the button is released. This behavior was broken
since the introduction of the scratch feature when the track is paused.
This commit solve this issue : the CUE point now acts like expected (CDJ
style), and we're still able to use the scratch feature when the track
is paused to easily navigate through the track.