trying to find the addresses to set the platter rotation speed precisely
The addresses to manage the platter aren't documented in the Denon
manual. I found some of them tho.

The platter rotation speed can be set to + x% from the normal RPM on
address `0x68` (104) and set to - x% from the normal RPM on address `0x69`
(105).
But these two addresses are changing the platter rotation speed on the
1% scale, which is not very precise.

I was looking for addresses to set the platter rotation speed on the
0.01% scale in order to have the precise pitch change applied to the
platter rotation speed.

Unfortunately, it does not seem to have such address :( .
Looks like it is not possible to set the platter rotation speed
precisely :( .