let the cue preview via DVS and prevent CUE tapping from moving the CUE point
Use the DVS sound signal to play the song while the CUE button is
pressed (i.e. do not tell mix to cue preview but use the DVS sound
signal sent by the SC3900 unit instead to preview the song).
Also prevent a little more the CUE point from moving when tapping the
CUE point for quick previews. As we're using DVS to play / preview the
song, the DVS sound signal is not as quite responsive as the CUE button
press itself (i.e. it takes time to the soundcard to detect that the
sound signal has changed). This time delta is now countered
programatically to try to remove it and to allow quick CUE previews w/o
moving the CUE point.