fix wrong hotcue light state when replacing a track on a deck
This commit fixes the following unexpected behavior :

- Load a track on deck A which has hotcue1 set. The SC3900 hotcue1
  button lights fully.
- Replace the track of deck A with another track which also has hotcue1
  set : the SC3900 hotcue1 button lights with backlight (dimmer) only,
  whereas it should light up fully as the hotcue1 is set in mixxx.

To fix this, we rely on the `hotcue_X_position` mixxx attribute instead
of the `hotcue_X_enabled` attribute. The position attribute has a `-1.0`
value when the hotcue is not set, so we can use this value with the `on`
and `off` XML config keys in order to send the correct value for the
button light (i.e. dimmer or fully on) to the SC3900.

See https://github.com/mixxxdj/mixxx/blob/release-2.2.4/src/controllers/midi/midioutputhandler.cpp#L43-L46

Notice that this unexpected behavior was only occuring for the first
hotcue light config exposed in the XML mapping (e.g. when the config for
the hotcue1 was removed, the behavior was occuring only to the hotcue2).
So this bug may also be related to how mixxx parses the mapping file,
but I have no clues about that.