let DVS signal handle playback on CUE press and release
Only use `cue_goto` instead of `cue_gotoandstop` for cursor navigation,
and let the DVS signal sent by the SC3900 unit be interpreted by the
soundcard to handle playback stop / start.

We still have to use a timer to handle DVS signal stop on CUE button
release as the MIDI signal for this release is interpreted by mixxx
before the stop of the DVS signal. So this timer allows us to make sure
we go back to the cuepoint once both signals are OK (MIDI signal for CUE
button release and the playback stop, i.e. silent DVS signal).